strong mode and type fixes fix deprecated info's in example
