+++ example/lib/material_app.dart
@@ -1,3 +1,4 @@
+import 'package:example/scroll_example.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_typeahead/flutter_typeahead.dart';
 
@@ -122,12 +123,12 @@ class _FormExampleState extends State<FormExample> {
             SizedBox(
               height: 10.0,
             ),
-            RaisedButton(
+            ElevatedButton(
               child: Text('Submit'),
               onPressed: () {
                 if (this._formKey.currentState!.validate()) {
                   this._formKey.currentState!.save();
-                  Scaffold.of(context).showSnackBar(
+                  ScaffoldMessenger.of(context).showSnackBar(
                     SnackBar(
                       content:
                           Text('Your Favorite City is ${this._selectedCity}'),
@@ -140,46 +141,6 @@ class _FormExampleState extends State<FormExample> {
         ),
       ),
     );
-  }
-}
-
-class ScrollExample extends StatelessWidget {
-  final List<String> items = List.generate(5, (index) => "Item $index");
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView(children: [
-      Center(
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Text("Suggestion box will resize when scrolling"),
-        ),
-      ),
-      SizedBox(height: 200),
-      TypeAheadField<String>(
-        getImmediateSuggestions: true,
-        textFieldConfiguration: TextFieldConfiguration(
-          decoration: InputDecoration(
-              border: OutlineInputBorder(),
-              hintText: 'What are you looking for?'),
-        ),
-        suggestionsCallback: (String pattern) async {
-          return items
-              .where((item) =>
-                  item.toLowerCase().startsWith(pattern.toLowerCase()))
-              .toList();
-        },
-        itemBuilder: (context, String suggestion) {
-          return ListTile(
-            title: Text(suggestion),
-          );
-        },
-        onSuggestionSelected: (String suggestion) {
-          print("Suggestion selected");
-        },
-      ),
-      SizedBox(height: 500),
-    ]);
   }
 }
