+++ example/lib/scroll_example.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_typeahead/flutter_typeahead.dart';
+
+class ScrollExample extends StatelessWidget {
+  final List<String> items = List.generate(5, (index) => "Item $index");
+
+  @override
+  Widget build(BuildContext context) {
+    return ListView(children: [
+      Center(
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Text("Suggestion box will resize when scrolling"),
+        ),
+      ),
+      SizedBox(height: 200),
+      TypeAheadField<String>(
+        getImmediateSuggestions: true,
+        textFieldConfiguration: TextFieldConfiguration(
+          decoration: InputDecoration(
+              border: OutlineInputBorder(),
+              hintText: 'What are you looking for?'),
+        ),
+        suggestionsCallback: (String pattern) async {
+          return items
+              .where((item) =>
+                  item.toLowerCase().startsWith(pattern.toLowerCase()))
+              .toList();
+        },
+        itemBuilder: (context, String suggestion) {
+          return ListTile(
+            title: Text(suggestion),
+          );
+        },
+        onSuggestionSelected: (String suggestion) {
+          print("Suggestion selected");
+        },
+      ),
+      SizedBox(height: 500),
+    ]);
+  }
+}
